requirements.txt
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
colorama==0.4.6
databases==0.7.0
DateTime==4.9
fastapi==0.89.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
psycopg2==2.9.5
pydantic==1.10.4
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
zope.interface==5.5.2
